refactor: split Comments component into subcomponents
